ADR-2336 | Scala 3 Upgrade (#30)
* ADR-2336 | Scala3 migration attempt

* ADR-2336 | Remove abstract from test classes

* ADR-2336 | Add comment to explain why models are in coverage exclude list for scala3

* ADR-2336 | Tidy up

* ADR-2336 | Remove unwanted scala style config

* ADR-2336 | Tidy up

* ADR-2336 | Fix PR bot coment

* Compile integration tests on run all checks

* ADR-2336 | bump versions

* ADR-2336 | Supress warnings

+++ app/uk/gov/hmrc/alcoholdutycontactpreferences/config/AppConfig.scala
@@ -58,7 +58,7 @@ class AppConfig @Inject() (config: Configuration, servicesConfig: ServicesConfig
   val enrolmentServiceName: String   = config.get[String]("enrolment.serviceName")
   val enrolmentIdentifierKey: String = config.get[String]("enrolment.identifierKey")
 
-  val dbTimeToLiveInSeconds: Int = 1200
+  val dbTimeToLiveInSeconds: Long = 1200
 
   def getSubscriptionUrl(appaId: String): String =
     s"$subscriptionHost$subscriptionGetSubscriptionUrlPrefix/$regime/$idType/$appaId"
